Bug 1324845 - Federation does not interoperate between MRG-M 2 and 3
An extra 'credit' argument was added to the link.bridge() management operation
in MRG3, which means the MRG2 qpid-route tool cannot add routes to a MRG3
broker.

The link.bridge() operation was deprecated in MRG2 in favour of
broker.create("bridge"...) but was still in use by the qpid-route tool.

The fix:
- Remove 'credit' argument from link.bridge() in new MRG3 broker.
- MRG2 qpid-route works with new MRG3 broker, same signature.
- New MRG3 qpid-route uses broker.create(), works with all brokers.
- Old MRG2 qpid-route uses link.brige() with optional credit, works with all brokers.